requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.4.1
astunparse==1.6.3
atari-py==0.2.6
attrs==19.3.0
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.20.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-colab==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.51.0
graphviz==0.14.1
grpcio==1.28.1
gym==0.17.2
h5py==2.10.0
httplib2==0.18.1
idna==2.8
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.21
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==5.7.8
numpy==1.18.3
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
portpicker==1.2.0
prometheus-client==0.8.0
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
PyDrive==1.3.1
pygame==1.9.6
pyglet==1.5.0
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-interface==1.6.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==4.5.3
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
